feat: create multi tree select
Introducing a method to give the chance for the user
to determine which of the multi values of a node tree
will be used. This is useful when we want to use just
the N last values whenever they're considered good enough,
for some arbitrary criteria
